feat: unit test for EspotiZeroconfServer localhost connection, get user info, and exception handling
